requirements.txt
# Core chatbot frameworks
Flask==2.2.5
gradio==3.44.0
gunicorn==21.2.0
flask-cors==4.0.0

# Document handling
python-docx==0.8.11
lxml==6.0.2

# Langchain and community extensions
langchain==0.2.12
langchain-community==0.2.11
langchain-huggingface==0.0.3

# Transformers & embeddings
transformers==4.44.2
sentence-transformers>=2.6.0,<3  # updated to satisfy langchain-huggingface

# Vector search
faiss-cpu==1.12.0
numpy>=1.26.4,<1.27

# Utilities
tqdm==4.66.1
requests==2.32.5
pydantic<3,>=1
PyYAML>=6.0.3
